Update to 1.2.0 Convert most environment variables to yaml config

+++ sharepyle/creds.py
@@ -7,6 +7,7 @@ import secrets
 import time
 import urllib.parse
 from pathlib import Path
+import yaml
 
 import requests
 from bs4 import BeautifulSoup
@@ -14,19 +15,32 @@ from http_requester.creds import UserCreds, Credentials
 
 logging.basicConfig(level=os.environ.get('LOGLEVEL', 'WARNING'))
 
-okta_api_key = os.environ.get('okta_api_key')
-client_id = os.environ.get('sharefile_client_id')
-client_secret = os.environ.get('sharefile_client_secret')
-sf_subdomain = os.environ.get('sf_subdomain')
-okta_subdomain = os.environ.get('okta_subdomain')
-app_name = 'sharefile'
-sf_okta_app_id = os.environ.get('sf_okta_app_id')
-okta_base_url = f"https://{okta_subdomain}.okta.com/api/v1"
 
-TOKEN_DIR = Path.home() / '.tokens'
+class MissingConfiguration(IOError):
+    pass
 
-BASE_URL = os.environ.get('sharefile_base_url')
-if not BASE_URL:
+
+SHAREPYLE_CONFIG_PATH = os.environ.get('SHAREPYLE_CONFIG_PATH', Path.home() / '.config' / 'sharepyle')
+if not SHAREPYLE_CONFIG_PATH:
+    raise MissingConfiguration('Missing local configuration path!')
+SHAREPYLE_CONFIG_FILE = SHAREPYLE_CONFIG_PATH / 'config.yaml'
+if not SHAREPYLE_CONFIG_FILE:
+    raise MissingConfiguration('Missing local configuration file!')
+
+with open(SHAREPYLE_CONFIG_FILE, 'r') as configfile:
+    SHAREPYLE_CONFIG = yaml.safe_load(configfile)
+
+SHAREFILE_CLIENT_ID = SHAREPYLE_CONFIG['sharefile']['client_id']
+SHAREFILE_CLIENT_SECRET = SHAREPYLE_CONFIG['sharefile']['client_secret']
+SHAREFILE_SUBDOMAIN = SHAREPYLE_CONFIG['sharefile']['subdomain']
+SHAREFILE_BASE_URL = SHAREPYLE_CONFIG['sharefile']['base_url']
+
+OKTA_API_KEY = SHAREPYLE_CONFIG['okta']['api_key']
+OKTA_SUBDOMAIN = SHAREPYLE_CONFIG['okta']['subdomain']
+OKTA_BASE_URL = SHAREPYLE_CONFIG['okta']['base_url']
+SHAREFILE_OKTA_APP_ID = SHAREPYLE_CONFIG['okta']['app_id']
+
+if not SHAREFILE_BASE_URL:
     raise AttributeError(f"No base url found in environment.")
 
 my_okta = UserCreds(
@@ -74,7 +88,7 @@ def sf_refresh(self):
 
 def get_okta_session_token(username: str, password: str, okta_api_key: str, session: requests.Session = None):
     session = session if session is not None else requests.Session()
-    okta_url = f"{okta_base_url}/authn"
+    okta_url = f"{OKTA_BASE_URL}/authn"
     headers = {
         'Authorization': f"SSWS {okta_api_key}",
         'Accept': 'application/json',
@@ -97,14 +111,14 @@ def get_sharefile_saml_request(client_id: str, session: requests.Session = None)
     session = session if session is not None else requests.Session()
     redirect_uri = 'https://secure.sharefile.com/oauth/oauthcomplete.aspx'
     state = secrets.token_urlsafe(128)
-    url = f"https://{sf_subdomain}.sharefile.com/saml/login"
+    url = f"https://{SHAREFILE_SUBDOMAIN}.sharefile.com/saml/login"
     params = {
         'response_type': 'code',
         'redirect_uri': redirect_uri,
         'client_id': client_id,
         'state': state,
         'oauth': 1,
-        'subdomain': sf_subdomain,
+        'subdomain': SHAREFILE_SUBDOMAIN,
         'appcp': 'sharefile.com',
         'apicp': 'sf-api.com'
     }
@@ -115,8 +129,8 @@ def get_sharefile_saml_request(client_id: str, session: requests.Session = None)
 
 def get_sharefile_saml_response(token: str, saml_request: str, okta_api_key: str, session: requests.Session = None):
     session = session if session is not None else requests.Session()
-    login_url = f"https://{okta_subdomain}.okta.com/login/sessionCookieRedirect"
-    redirect_url_base = f"https://{okta_subdomain}.okta.com/app/sharefile/{sf_okta_app_id}/sso/saml"
+    login_url = f"https://{OKTA_SUBDOMAIN}.okta.com/login/sessionCookieRedirect"
+    redirect_url_base = f"https://{OKTA_SUBDOMAIN}.okta.com/app/sharefile/{SHAREFILE_OKTA_APP_ID}/sso/saml"
     redirect_url = f"{redirect_url_base}?SAMLRequest={urllib.parse.quote(saml_request)}"
     headers = {
         'Authorization': f"SSWS {okta_api_key}",
@@ -139,7 +153,7 @@ def get_sharefile_auth_code(saml_response: str, session: requests.Session = None
         'SAMLResponse': saml_response
     }
     headers = {'Content-Type': 'application/x-www-form-urlencoded'}
-    url = f"https://{sf_subdomain}.sharefile.com/saml/acs"
+    url = f"https://{SHAREFILE_SUBDOMAIN}.sharefile.com/saml/acs"
     t = session.post(url, headers=headers, data=data)
     location = t.history[1].headers.get('location')
     return location.split('?', 1)[1].split('=', 1)[1].split('&', 1)[0], session
@@ -147,7 +161,7 @@ def get_sharefile_auth_code(saml_response: str, session: requests.Session = None
 
 def get_sharefile_access_tokens(code: str, client_id: str, client_secret: str, session: requests.Session = None):
     session = session if session is not None else requests.Session()
-    oauth_url = f"https://{sf_subdomain}.sf-api.com/oauth/token"
+    oauth_url = f"https://{SHAREFILE_SUBDOMAIN}.sf-api.com/oauth/token"
     redirect_uri = 'https://secure.sharefile.com/oauth/oauthcomplete.aspx'
     auth_string = f"{client_id}:{client_secret}"
     headers = {
@@ -170,16 +184,18 @@ def get_sharefile_access_tokens(code: str, client_id: str, client_secret: str, s
 
 
 def get_sharefile_credentials(
-        session: requests.Session = None
+        session: requests.Session = None,
+        force_refresh: bool = True
 ) -> Credentials:
     """Shows basic usage of the People API.
     Prints the name of the first 10 connections.
     """
     creds = None
-    token_path = TOKEN_DIR / 'sftoken.pickle'
-    if os.path.exists(token_path):
-        with open(token_path, 'rb') as token:
-            creds = pickle.load(token)
+    if not force_refresh:
+        token_path = TOKEN_DIR / 'sftoken.pickle'
+        if os.path.exists(token_path):
+            with open(token_path, 'rb') as token:
+                creds = pickle.load(token)
     if not creds or not creds.valid:
         if creds and creds.expired and creds.refresh_token:
             creds.refresh()
@@ -190,25 +206,26 @@ def get_sharefile_credentials(
                 'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.105 Safari/537.36'
             })
             '''
-            token, session = get_okta_session_token(my_okta.email, my_okta.password, okta_api_key, session)
-            saml_request, session = get_sharefile_saml_request(client_id, session)
-            saml_response, session = get_sharefile_saml_response(token, saml_request, okta_api_key, session)
+            token, session = get_okta_session_token(my_okta.email, my_okta.password, OKTA_API_KEY, session)
+            saml_request, session = get_sharefile_saml_request(SHAREFILE_CLIENT_ID, session)
+            saml_response, session = get_sharefile_saml_response(token, saml_request, OKTA_API_KEY, session)
             code, session = get_sharefile_auth_code(saml_response, session)
-            access_token, refresh_token, expiration = get_sharefile_access_tokens(code, client_id,
-                                                                                  client_secret,
+            access_token, refresh_token, expiration = get_sharefile_access_tokens(code, SHAREFILE_CLIENT_ID,
+                                                                                  SHAREFILE_CLIENT_SECRET,
                                                                                   session)
             creds = Credentials(
                 token=access_token,
                 refresh_token=refresh_token,
                 expiration=expiration,
-                client_id=client_id,
-                client_secret=client_secret,
-                token_url=BASE_URL,
+                client_id=SHAREFILE_CLIENT_ID,
+                client_secret=SHAREFILE_CLIENT_SECRET,
+                token_url=SHAREFILE_BASE_URL,
                 format_matrix=(
                     ('Authorization', ('Bearer {}', 'token')),
                 ),
                 refresh_func=sf_refresh
             )
-        with open(token_path, 'wb') as token:
-            pickle.dump(creds, token)
+        if not force_refresh:
+            with open(token_path, 'wb') as token:
+                pickle.dump(creds, token)
     return creds
